KEIJO-157: Separate option logic to custom hook
